refactor: split form-builder and field-builder into separate files
