+++ packages/@tinacms/form-builder/src/Form.tsx
@@ -20,7 +20,7 @@ import * as React from 'react'
 import { useState, useContext } from 'react'
 import PropTypes from 'prop-types'
 import { Form } from '@tinacms/forms'
-import { FormBuilder } from './final-form-builder'
+import { FormBuilder } from './form-builder'
 import { Field } from 'react-final-form'
 
 interface RenderProps {
