+++ packages/@tinacms/form-builder/src/index.ts
@@ -17,5 +17,6 @@ limitations under the License.
 */
 
 export * from './field-plugin'
-export * from './final-form-builder'
+export * from './form-builder'
+export * from './fields-builder'
 export * from './Form'
